Change navigation disabled behavior
The navigation buttons that move the weeks forward and back are no
correctly disabled when you reach the bounds of the calendar. If they are
passed in as buttons they are given the "disabled" attribute set to
disabled, or given no disabled attribute. If they are passed in as anotehr
element type they are given the aria-disabled="true" attribute and the
aria-disabled="false" if they are not enabled.