Add selected date range property
After selecting a range of times, the public property "selected" will be
updated to hold an object containing the start and end date time object.
This is set when the selectFinish event is triggered. This allows the user
to get the latest range by binding the the selectFinish event. Remvoed
unit tests for testing user interaction as they were not behaving
properly.